[IMP] l10n_es_facturae: Clean XML file when it is not signed
