+++ l10n_es_facturae/reports/__init__.py
@@ -1,1 +1,2 @@
-from . import report_facturae
+from . import report_template_facturae
+from . import report_facturae_signed
